django_events/requirements.txt
django
djangorestframework
gunicorn